provide a more seamless experience with dev docker

+++ docker/dev/run.sh
@@ -8,6 +8,14 @@ git pull
 # change to code folder
 cd /code/
 
-pip install -e .
+if [ -f "setup.py" ]; then
+    pip install -e .
+fi
 
+# start a new tmux session
+tmux new -d
+tmux send-keys "nvim" Enter
+tmux split-window -v -p 10
+
+# start a new terminal, in case runs are started from a tmux session
 fish
